# whatsapp-clone-ui/src/app/services/models/chat-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ChatResponse {
  chatId?: string;
  lastMessage?: string;
  lastMessageTime?: string;
  name?: string;
  receiverId?: string;
  receiverOnline?: boolean;
  senderId?: string;
  unreadCount?: number;
}