# whatsappclone/src/main/java/com/thm/whatsappclone/request/MessageRequest.java
package com.thm.whatsappclone.request;


import com.thm.whatsappclone.message.MessageType;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class MessageRequest {

    private String content;
    private String senderId;
    private String receiverId;
    private MessageType type;
    private String chatId;

}