# whatsappclone/src/main/java/com/thm/whatsappclone/response/MessageResponse.java
package com.thm.whatsappclone.response;

import com.thm.whatsappclone.message.MessageState;
import com.thm.whatsappclone.message.MessageType;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class MessageResponse {

    private Long id;
    private String content;
    private MessageType type;
    private MessageState state;
    private String senderId;
    private String receiverId;
    private LocalDateTime createdAt;
    private byte[] media;
}