# whatsappclone/src/main/java/com/thm/whatsappclone/response/StringResponse.java
package com.thm.whatsappclone.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class StringResponse {
    private String response;
}